Feature/cms2

+++ app/controllers/admin/articles_controller.rb
@@ -20,11 +20,14 @@ module Admin
 
     def edit
       @article = Article.find(params[:id])
+      raise ActiveRecord::RecordNotFound  unless @article.user_id == current_user.id
     end
 
     def update
       @article = Article.find(params[:id])
-      if @article.update(article_params)
+      raise ActiveRecord::RecordNotFound  unless @article.user_id == current_user.id
+      @article.attributes = article_params
+      if @article.update_with_history!(current_user)
         redirect_to articles_path, notice: "記事を更新しました"
       else
         render :edit
@@ -33,6 +36,7 @@ module Admin
 
     def destroy
       @article = Article.find(params[:id])
+      raise ActiveRecord::RecordNotFound  unless @article.user_id == current_user.id
       if @article.destroy
         redirect_to articles_path, notice: "記事を削除しました"
       else
